Create initial CP2 query for Contacts Picker UI
It replaces a hardcoded list of dummy contacts with a query to the on device contacts.

It is not a final version of the query, it will be revisited with a short design doc and a review to ensure the result of the query is as desired.

Screenshot of the multiple emails per one contacts:
http://screen/6cYxpoY5L8z4oez.png For now multiple emails/phones are
concatented in a comma separated string. Support for drop down and
multiple email/phone entries will be done in a follow up.

Bug: 442966559
Test: atest ContactsPickerIntegrationTest ContactsPickerComponentTests
Flag: android.content.flags.enable_system_contacts_picker
Change-Id: I49c0573e0a9901679f4db214383debb676a9e11d