Change Contact to a sealed class for display mode
This change refactors the way the display mode is handled in the Contacts Picker. It replaces the DisplayMode enum and DisplayModeResolver with a Contact sealed class.

This centralises the logic for storing the display mode and improves type safety, as it now reacts to the type of the contact object (`BasicContact`, `PhoneContact`, or `EmailContact`) to determine what to display.

It introduces the ContactsRepository which for now has hardcoded values
but will be changed to queries to CP2 for real on device contacts.

This is a refactoring with no UI visible change.

Bug: 436818961
Test: atest ContactsPickerIntegrationTest ContactsPickerComponentTests
Flag: android.content.flags.enable_system_contacts_picker
Change-Id: I096669933b76aeaa1e2f65cd57460b894e2c5a10