ACTION_PICK takeover for ContactsPicker
Raise the priority of the ContactsPickerActivity intent filter for
ACTION_PICK for contacts related mime types. This makes all the intents
be intercepted by the CPA.

After checking the targetSdk of the calling app the CPA decides to
handle the intent within the new picker app for apps targeting SDKs 37+
and re-sends it to other handlers of that same intent for apps targeting
SDKs below 37.

For now using the ChooserIntent which will be revisited to ideally use
the ResolverActivity to let the user choose the app to handle the intent
for apps targeting SDKs below 37.

Bug: 441477608
Test: atest ContactsPickerIntegrationTest ContactsPickerComponentTests
Flag: android.content.flags.enable_system_contacts_picker
Change-Id: Ia1992054cc358c8e09d7840d6f57f93898fc2086