Integrate Compose Navigation for screen routing
This commit introduces the Compose Navigation library to handle navigation between the ContactsPicker and PrivacyDetails screens.
This is a non-UI change that sets up the navigation graph.

Key changes include:
- Defining routes for each screen.
- Implementing the logic for navigating between them.

Note: The integration of the bottom sheet will be handled in a follow-up CL.

Bug: 446119760
Test: atest ContactsPickerIntegrationTests ContactsPickerComponentTests
Flag: android.content.flags.enable_system_contacts_picker
Change-Id: I01042e454fa3b10de24742cead252877c2ee601c