Remove setting the sheetShape
Setting the sheetShape makes all 4 corners rounded but the bottom
corners are not supposed to be rounded to give the effect of an
"infinite sheet" even when fully expanded to the top of the screen. Since the material theme already correctly sets the
top corners to the desired shape we can just remove the param.

Before the change: http://screen/5mFHC4mVfeDwgsJ.png
After the change: http://screen/3yJCUmaTaNjxDtK.png

Test: manually
Bug: 436818961
Flag: android.content.flags.enable_system_contacts_picker
Change-Id: If0a1aeb904f78a0a0b02a08aa5ee0b348e4749f3