Pass the state of ContactsUiState instead of value
Don't read the state of the ContactsUiState in the
ContactsPickerActivity and pass the state instead all the way down to
the first composable that needs to react to changes of the
ContactsUiState.

This will reduce the recompositions of the composables that only take
the state and do not read it.

Bug: 436818961
Test: atest ContactsPickerIntegrationTests ContactsPickerComponentTests
Flag: android.content.flags.enable_system_contacts_picker
Change-Id: Ie193974b505f6f97f814e35ba76e3aace1567071