Pass selected Uris as results to the calling app
When the "Done" button is clicked on the selection bar then create the
Uris to the selected contacts and pass them in the setActivityResult.

For EmailContact and PhoneContact build the Uri with the Id from the
Data table, for the DisplayNameContact we return the whole contact so
use the stable lookupKey.

Bug: 452020365, 441478493
Test: atest ContactsPickerIntegrationTests ContactsPickerComponentTests
Flag: android.content.flags.enable_system_contacts_picker

Change-Id: I53e5fb282ceed082197943523b18e4919baee9d2