Add Contacts Picker client apps
There are two apps, targeting API 36 and 37 to test the ACTION_PICK
takeover. The apps reuse the logic to prepare and send the ACTION_PICK
intent, but the logic to use ACTION_PICK_CONTACTS is defined and used
only in the SDK 37 app since the API is SDK guarded. The UIs look almost the same with the sub title indicating the
target SDK of the client app. The icons and names state 36 and 37 to
make it easier to use.

The modules are never intended to be productionalized.

Video of using each one in sequence:
https://drive.google.com/file/d/1YZe7q6a5yVEYfwfEpfjTAZyNP5ZHU12P/view?usp=sharing&resourcekey=0-Oi-1Di5Npe-1kXktXBhEFQ
(looks like Google Contacts is blocking screenrecord? the black screen
is when GC is in foreground).

Bug: 447114080
Test: N/A
Flag: EXEMPT TEST_ONLY
Change-Id: I511b77fc976eaf383a94983a79cb365630f2e6f4