Add support for EXTRA_USE_SYSTEM_CONTACTS_PICKER
Change the ContactsPickerActivity to route to itself when the extra is
set.

Additionally add a toggle to the demo client app SDK 36 to set the
extra and display it only when the contacts picker flag is enabled.

Video shows a different behaviour in the new toggle off and on as well
as no toggle in the SDK 37 demo app: https://drive.google.com/file/d/1MtQXEvkFVGpzka9T13gPTDSFMzWrJib5/view?usp=sharing&resourcekey=0-LK-haWKsGFfRlX-Qvpp2ig
Bug: 447114080
Test: atest ContactsPickerIntegrationTests ContactsPickerComponentTests
Flag: android.content.flags.enable_system_contacts_picker
Change-Id: I56163de4d3f84758578b722313e4904c10ab865c

Change-Id: I9eb5cb511a4120548c4f115a8aec4e92708fe4d5