Add maximum allowed contacts selection
By default allow selection of 50 contacts in multi select mode.
Check the intent for EXTRA_PICK_CONTACTS_SELECTION_LIMIT if the multi
select is enabled and allow overriding the default to up to 100.

When the limit is reached display a snackbar when a user
tries to select more contacts.

Screenshot of the snackbar: http://screen/6XMHeWMXsNv68of.png

Bug: 442401361
Test: atest ContactsPickerIntegrationTests ContactsPickerComponentTests
Flag: android.content.flags.enable_system_contacts_picker
Change-Id: I80424f1ff7c22fe131fddfbe08e8aa8bc58a9e04