Add flag guarded takeover of ACTION_PICK for contacts
Add a flag declaration in the ContactsPicker repo and if enabled always
handle the ACTION_PICK internally in the Contacts Picker.

The flag is intended to only ever reach Trunkfood and be used to test
the new Contacts Picker in all Trunkfood population.

Bug: 459413010
Test: atest ContactsPickerIntegrationTests ContactsPickerComponentTests
Flag: android.content.flags.enable_system_contacts_picker
Change-Id: Iadefdf595c5ccd01f66f93203535c6826dce2c2c