Remove the duplicate resource string
Removing was blocked on failing integration tests but it was caused by a
wrong setup of the tests. Adding ContactsPickerLib to the associates of
the instrumentation tests was going to fail sooner or later since the
resources were build in two ways.

Fix: 459517167
Test: atest ContactsPickerComponentTests ContactsPickerIntegrationTests
Flag: android.content.flags.enable_system_contacts_picker
Change-Id: I2d8165e687c24f4c40fa2e4e44696ded8957a576