Add ContactsPickerRequestConfig and ContactsQueryMode
The new data class is derived from the values passed in the incoming
intent and is used to control the mode of querying contacts. Also moving
the logic to parse the intent from the ContactsViewModel to the new
ContactsPickerRequestConfig companion object.

This change adds partial support for the ACTION_PICK_CONTACTS but it is not yet declared in the manifest so only adding some unit tests.

Side change: remove parametrized testing from the
ContactsRepositoryImplTest. Now that the Repository is not responsible
for the intent action and type parsing it is simpler and more
straightforward to just have the tests for all the enum cases. Not to
mention that the parametrized tests were set up wrong and running
unnecessary iterations of some tests

Bug: 452020367
Test: atest ContactsPickerIntegrationTests ContactsPickerComponentTests
Flag: android.content.flags.enable_system_contacts_picker
Change-Id: I3ff9fcfbda3484ccdf1bba1988dcd83500a6dc2f