Remove SetActivityResult
The logic to create the return session Uri for the ACTION_PICK_CONTACTS
will be handled in its own repository so SetActivityResult is redundant.
Instead move the fairly simple logic to create the return Intent for the
ACTION_PICK directly to the ContactsViewModel.

Bug: 452020367
Test: atest ContactsPickerIntegrationTests ContactsPickerComponentTests
Flag: android.content.flags.enable_system_contacts_picker
Change-Id: I5a1bf17514d063e6b2e843ea72ebea1e139ba34b