Change the queried URI to ContactsContract.AUTHORITY_URI for custom
There is a mismatch in the matcher used in CP2 and the query build in
the ContactsRepositoryImpl and it results in an error. This fix makes
the query resolve correctly. Added a TODO to consider a more appropriate
fix in a follow up.

Bug: 467326511
Test: atest ContactsPickerIntegrationTests ContactsPickerComponentTests
Flag: android.content.flags.enable_system_contacts_picker
Change-Id: Ic98851e1936419876ace9ea75004a8638038dccb