Add querying for specific data rows for custom ACTION_PICK_CONTACTS
The new additional query is executed after the user completes the
selection of the contacts and presses the "Done" button. In the contacts
picker only the reference to the contact IDs are kept to save the memory
pressure. After the user finalises their choice, query for all requested
mime type data rows for the selected contacts.

Bug: 452020367
Test: atest ContactsPickerIntegrationTests ContactsPickerComponentTests
Flag: android.content.flags.enable_system_contacts_picker
Change-Id: I090f2ef0fbb69ced13232e40283199114493fbc0